[Quant] Separated implementations for quantized & non-quantized tensors for the unsqueeze function
Summary: This is a PR that addresses a concern brought up in https://github.com/pytorch/pytorch/issues/54150 regarding opting for separat kernels for quantized & non-quantized tensors, when possible, as opposed to using conditional if-else.

Test plan: Additional testing was not implemented to test this change because test cases in the existing test suite already make use of the squeeze function for various backends.

ghstack-source-id: 02486bec3ae94f3695c78639d20915c01744a0a4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71648